+++ components/Cms.php
@@ -7,6 +7,7 @@
  */
 namespace skeeks\cms\components;
 
+use skeeks\cms\assets\CmsAsset;
 use skeeks\cms\base\components\Descriptor;
 use skeeks\cms\base\db\ActiveRecord;
 use skeeks\cms\base\Module;
@@ -136,7 +137,7 @@ class Cms extends \skeeks\cms\base\Component
     /**
      * @var string Это изображение показывается в тех случаях, когда не найдено основное.
      */
-    public $noImageUrl          = 'http://vk.com/images/deactivated_100.gif';
+    public $noImageUrl;
 
     /**
      * @var array
@@ -269,6 +270,11 @@ class Cms extends \skeeks\cms\base\Component
     public function init()
     {
         parent::init();
+
+        if (!$this->noImageUrl)
+        {
+            $this->noImageUrl = CmsAsset::getAssetUrl('img/image-not-found.jpg');
+        }
 
         //Название проекта.
         if (!$this->appName)